requirements.txt
# Core dependencies
chainlit==0.7.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.0

# API and utilities
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
httpx==0.25.0

# Testing
pytest==7.4.3

# Logging
loguru==0.7.2
